requirements.txt
python-telegram-bot
yfinance
pandas
pandas-ta
gradio